add basic filters

+++ lakesoul-presto/src/main/java/com/facebook/presto/lakesoul/util/PrestoUtil.java
@@ -123,7 +123,8 @@ public class PrestoUtil {
             if(type.getPrecision() == FloatingPointPrecision.HALF){
                 return UnknownType.UNKNOWN;
             }else if(type.getPrecision() == FloatingPointPrecision.SINGLE){
-                return FloatType.FLOAT;
+                //return FloatType.FLOAT;
+                return DoubleType.DOUBLE;
             }else if(type.getPrecision() == FloatingPointPrecision.DOUBLE) {
                 return DoubleType.DOUBLE;
             }
